% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bedToGRanges}
\alias{bedToGRanges}
\title{Import TSS bed file as GRanges}
\usage{
bedToGRanges(bed)
}
\arguments{
\item{bed}{BedFile to import}
}
\value{
GRanges object
}
\description{
Import TSS bed file as GRanges
}

