% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cager_wrapper.R
\name{cage_wrapper}
\alias{cage_wrapper}
\title{A wrapper for CAGE Normalization and TSS calling}
\usage{
cage_wrapper(input, labels, ncores = NULL, tpmCutoff = 10,
  genome = "BSgenome.Dmelanogaster.UCSC.dm6")
}
\arguments{
\item{input}{path to input bam file(s)}

\item{labels}{sample labels}

\item{ncores}{number of threads (leave NULL for one)}

\item{tpmCutoff}{TPM cutoff for TSS calling}

\item{genome}{The name of corresponding BSGenome object (must be installed)}
}
\value{
QC plots, Normalized bedgraph files and TSS bed files
}
\description{
A wrapper for CAGE Normalization and TSS calling
}
\examples{

cager_wrapper(input, labels, ncores = NULL, tpmCutoff = 10, genome = "BSgenome.Dmelanogaster.UCSC.dm6")

}

