% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDups.R
\name{filterDuplicates}
\alias{filterDuplicates}
\title{Filter PCR-duplicates from BAM file using internal UMIs}
\usage{
filterDuplicates(bamFile, outFile)
}
\arguments{
\item{bamFile}{Input BAM file}

\item{outFile}{Output BAM file}
}
\value{
Filtered BAM file (with only R1), after PCR duplicate removal
}
\description{
Filter PCR-duplicates from BAM file using internal UMIs
}
\examples{

filterDuplicates(bamFile = "FATSCapR/tests/testthat/test_sorted.bam",
		outFile = "FATSCapR/tests/testthat/test_pcrRem.bam")
}

