% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFETISH.R
\name{mapFETISH}
\alias{mapFETISH}
\title{Map the paired-end FETISH data.}
\usage{
mapFETISH(index, R1, R2, output, nthreads, PE_orientation = "rf", ...)
}
\arguments{
\item{index}{Subread Index file for the genome}

\item{R1}{forward read (_R1) fastq.}

\item{R2}{reverse read (_R2) fastq.}

\item{output}{output file name}

\item{nthreads}{number of threads to use for mapping.}

\item{PE_orientation}{library type.}
}
\value{
bamfile A mapped BAM file for the sample.
}
\description{
Map the paired-end FETISH data.
}
\examples{


}

