% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFETISH.R
\name{split_BAM}
\alias{split_BAM}
\title{Split the composite BAM file using internal indexes}
\usage{
split_BAM(bamFile, index_list, outfile_list, rep_indexList = NULL, nthreads)
}
\arguments{
\item{bamFile}{Input BAM file}

\item{index_list}{A list of indexes for splitting (character vector)}

\item{outfile_list}{A list of output file names (with order corresponding to that of index_list)}

\item{rep_indexList}{A list of replicate indexes (if you want to split by the replicates too)}

\item{nthreads}{Number of threads}
}
\value{
Filtured files
}
\description{
Split the composite BAM file using internal indexes
}
\examples{
filter_BAM(bamFile = "FATSCapR/tests/testthat/test_sorted.bam",
 	   index_list = c("TTAGCC" ,"CAAGTG"),
 	    	   outfile_list = c("FATSCapR/tests/testthat/test_filt1.bam",
 	    	   "FATSCapR/tests/testthat/test_filt2.bam"), nthreads = 10)
}

