% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBAM_byIndex.R
\name{splitBAM_byIndex}
\alias{splitBAM_byIndex}
\title{Split the composite BAM file using internal indexes}
\usage{
splitBAM_byIndex(bamFile, index_list, outfile_list, max_mismatch = 0,
  nthreads = 1)
}
\arguments{
\item{bamFile}{Input BAM file}

\item{index_list}{A list of indexes for splitting (character vector)}

\item{outfile_list}{A list of output file names (with order corresponding to that of index_list)}

\item{max_mismatch}{No. of mismatches allowed in index (maxium 1 recommended)}

\item{nthreads}{Number of threads}
}
\value{
Filtered files
}
\description{
Split the composite BAM file using internal indexes
}
\examples{
bam <- system.file("extdata", "test_mapped.bam", package = "FATSCapR")
splitBAM_byIndex(bamFile = bam,
 	   index_list = c("TTAGCC" ,"CAAGTG"),
 	    	   outfile_list = c("test_filt1.bam","test_filt2.bam"), nthreads = 10)
}

