% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFETISH.R
\name{trimFETISH}
\alias{trimFETISH}
\title{Trim the R2 of FETISH data and tag the read headers}
\usage{
trimFETISH(infile, outfile)
}
\arguments{
\item{infile}{Input _R2.fastq (or _R2.fastq.gz) file.}

\item{outfile}{Name of the Output fastq file.}
}
\value{
trimmedFile : The trimmed R2.fastq file.
}
\description{
Trim the R2 of FETISH data and tag the read headers
}
\examples{
trimFETISH("testdata/test_R2.fastq.gz","test_trimmed_R2.fastq")

}

