% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cager_wrapper.R
\name{cage_wrapper}
\alias{cage_wrapper}
\title{A wrapper for CAGE Normalization and TSS calling}
\usage{
cage_wrapper(input, labels, ncores = NULL, tpmCutoff = 10,
  genome = "BSgenome.Dmelanogaster.UCSC.dm6", find_TSSshift = FALSE,
  scoreshift_groupX, scoreshift_groupY, promoterShift_outFile)
}
\arguments{
\item{input}{path to input bam file(s)}

\item{labels}{sample labels}

\item{ncores}{number of threads (leave NULL for one)}

\item{tpmCutoff}{TPM cutoff for TSS calling}

\item{genome}{The name of corresponding BSGenome object (must be installed)}

\item{find_TSSshift}{logical, indicating if TSS shift between two samples are to be calculated}

\item{scoreshift_groupX}{Test sample for TSS shift}

\item{scoreshift_groupY}{Control sample for TSS shift}

\item{promoterShift_outFile}{Output file name for calculated TSS shift positions}
}
\value{
QC plots, Normalized bedgraph files and TSS bed files
}
\description{
A wrapper for CAGE Normalization and TSS calling
}
\examples{

bam <- system.file("extdata", "test_mapped.bam", package = "FATSCapR")
cager_wrapper(input = bam, labels = "test", tpmCutoff = 10, genome = "BSgenome.Dmelanogaster.UCSC.dm6")

}

