% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFETISH.R
\name{mapFETISH}
\alias{mapFETISH}
\title{Map the paired-end FETISH data.}
\usage{
mapFETISH(index, R1, R2, output, nthreads, logfile = NULL, ...)
}
\arguments{
\item{index}{character string giving the basename of the Subread index file.
Index files should be located in the current directory.}

\item{R1}{forward read (_R1) fastq.}

\item{R2}{reverse read (_R2) fastq.}

\item{output}{output file name}

\item{nthreads}{number of threads to use for mapping.}

\item{logfile}{a log file to write the processing message.}

\item{...}{additional arguments passed to the RSubread::align function.}
}
\value{
bamfile A mapped BAM file for the sample.
}
\description{
Map the paired-end FETISH data.
}
\examples{
mapFETISH(index,R1,R2,output,nthreads,logfile=NULL)

}

